chore: Move ConversationSessionResetSystemMessageCellDescription into its own file
